requirements.txt
nb-cli
nonebot-plugin-apscheduler
nonebot-adapter-onebot
xmltodict
pypinyin
Pillow
psutil
# ---------------------------------
# add by Judy on 13rd July 2021
# tushare abandoned on 17th July 2021
# pandas~=1.3.3
# ----------------------------------
# add by Judy on 17 July 2021
baostock
pydantic
httpx
nonebot2[websockets]
